+++ src/main/java/fr/maxlego08/menu/ZItemManager.java
@@ -4,6 +4,7 @@ import fr.maxlego08.menu.api.ItemManager;
 import fr.maxlego08.menu.api.MenuItemStack;
 import fr.maxlego08.menu.api.MenuPlugin;
 import fr.maxlego08.menu.api.configuration.Config;
+import fr.maxlego08.menu.api.event.events.ZMenuItemsLoad;
 import fr.maxlego08.menu.api.mechanic.MechanicFactory;
 import fr.maxlego08.menu.api.mechanic.MechanicListener;
 import fr.maxlego08.menu.item.CustomItemData;
@@ -17,6 +18,7 @@ import org.bukkit.event.HandlerList;
 import org.bukkit.event.Listener;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
+import org.bukkit.persistence.PersistentDataContainer;
 import org.bukkit.persistence.PersistentDataType;
 import org.bukkit.plugin.Plugin;
 
@@ -34,9 +36,11 @@ public class ZItemManager implements ItemManager{
     private final MenuPlugin menuPlugin;
 
     private final Map<String, MechanicListener> mechanicListeners = new HashMap<>();
-    private final Map<String, MechanicFactory> mechanicFactories = new HashMap<>();
+    private final Map<String, MechanicFactory<?>> mechanicFactories = new HashMap<>();
 
     private final Map<String, CustomItemData> customItems = new HashMap<>();
+
+    private boolean isFirstLoad = true;
 
     public ZItemManager(ZMenuPlugin menuPlugin) {
         this.itemIdKey = new NamespacedKey(menuPlugin, "item-id");
@@ -66,6 +70,11 @@ public class ZItemManager implements ItemManager{
         }
         try (Stream<Path> stream = Files.walk(Paths.get(itemsFolder.getPath()))) {
             stream.skip(1).map(Path::toFile).filter(File::isFile).filter(e -> e.getName().endsWith(".yml")).forEach(this::loadCustomItem);
+            ZMenuItemsLoad event = new ZMenuItemsLoad(new HashSet<>(customItems.keySet()), !isFirstLoad);
+            menuPlugin.getServer().getPluginManager().callEvent(event);
+            if (isFirstLoad) {
+                isFirstLoad = false;
+            }
         } catch (IOException exception) {
             if (Config.enableDebug){
                 Logger.info("Error while loading items: " + exception.getMessage(), Logger.LogType.ERROR);
@@ -86,7 +95,7 @@ public class ZItemManager implements ItemManager{
                 if (mechanicSection != null) {
                     path += "mechanics.";
                     for (String mechanicId : mechanicSection.getKeys(false)) {
-                        MechanicFactory factory = mechanicFactories.get(mechanicId);
+                        MechanicFactory<?> factory = mechanicFactories.get(mechanicId);
                         if (factory != null) {
                             factory.parse(this.menuPlugin, itemId, mechanicSection.getConfigurationSection(mechanicId), config, file, path + mechanicId + ".");
                             mechanicIds.add(mechanicId);
@@ -108,7 +117,7 @@ public class ZItemManager implements ItemManager{
     @Override
     public void reloadCustomItems() {
         customItems.clear();
-        for (MechanicFactory factory : mechanicFactories.values()) {
+        for (MechanicFactory<?> factory : mechanicFactories.values()) {
             factory.clearMechanics();
         }
         this.loadCustomItems();
@@ -158,7 +167,7 @@ public class ZItemManager implements ItemManager{
     }
 
     @Override
-    public void registerMechanicFactory(MechanicFactory factory) {
+    public void registerMechanicFactory(MechanicFactory<?> factory) {
         if (mechanicFactories.containsKey(factory.getMechanicId())) {
             Logger.info("MechanicFactory " + factory.getMechanicId() + " is already registered.", Logger.LogType.WARNING);
             return;
@@ -177,9 +186,12 @@ public class ZItemManager implements ItemManager{
         MenuItemStack menuItemStack = itemData.menuItemStack();
         ItemStack itemStack = menuItemStack.build(player);
         ItemMeta itemMeta = itemStack.getItemMeta();
-        itemMeta.getPersistentDataContainer().set(itemIdKey, PersistentDataType.STRING, itemId);
-        itemMeta.getPersistentDataContainer().set(ownerKey, PersistentDataType.STRING, player.getUniqueId().toString());
-        itemStack.setItemMeta(itemMeta);
+        if (itemMeta != null) {
+            PersistentDataContainer persistentDataContainer = itemMeta.getPersistentDataContainer();
+            persistentDataContainer.set(itemIdKey, PersistentDataType.STRING, itemId);
+            persistentDataContainer.set(ownerKey, PersistentDataType.STRING, player.getUniqueId().toString());
+            itemStack.setItemMeta(itemMeta);
+        }
 
         boolean shouldCancel = false;
         for (String mechanicId : itemData.mechanicIds()) {