+++ changelog.md
@@ -42,6 +42,21 @@
 
 # Unreleased
 
+# 1.1.0.7
+
+- Fixed NMS compatibility for **1.21.10**.
+- Fixed issues with **DeluxeMenus** inventory loading.
+- Fixed **MiniMessage** regex handling.
+- Added the ``ZMenuItemsLoad`` event, triggered when zMenu custom items are loaded or reloaded.
+- Improved the logging system.
+- Improved item mechanics.
+- Added ``deny-chance-actions`` support for actions.
+- Various minor fixes and general code improvements.
+- Added new placeholders:  
+  ``%zmenu_time_unix_timestamp%``,  
+  ``%zmenu_time_next_day_unix_timestamp%``,  
+  ``%zmenu_time_today_start_unix_timestamp%``
+
 # 1.1.0.6
 
 - Fixed the removal of default attributes.